feat(food-details): added localization and make widgets in files .
