+++ lib/shared_layers/localization/l10n_arb_files/app_ar.arb
@@ -91,7 +91,11 @@
   "explore": "استكشاف",
   "aiChat": "دردشة AI",
   "workouts": "التمارين",
-  "profile": "الملف الشخصي"
+  "profile": "الملف الشخصي",
+  "ingredients": "المكونات",
+  "recommendation": "التوصية",
+  "energy": "الطاقة",
+  "energyValue": "١٠٠ ك"
 
 
 }
